requirements.txt
absl-py==2.3.1
astunparse==1.6.3
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
Flask==3.1.2
flatbuffers==25.12.19
gast==0.7.0
google-pasta==0.2.0
grpcio==1.76.0
h5py==3.15.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
keras==3.13.0
libclang==18.1.1
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
namex==0.1.0
numpy==2.4.0
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pillow==12.0.0
protobuf==6.33.2
Pygments==2.19.2
requests==2.32.5
rich==14.2.0
setuptools==80.9.0
six==1.17.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
tf_keras==2.20.1
typing_extensions==4.15.0
urllib3==2.6.2
Werkzeug==3.1.4
wheel==0.45.1
wrapt==2.0.1
